+++ VersionHistory.md
@@ -6,10 +6,12 @@ Describe changes here when they're committed to the `master` branch. Move them t
 
 Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in accordance with [Semantic Versioning](https://semver.org/).
 
+## Released
+
+### 3.6.0
+
 * [patch] Run nuget.exe in non-interactive mode.
 * [minor] Allow customization of how tests are run.
-
-## Released
 
 ### 3.5.0